chore(loadouts): delete unused edit form, move create/upload loadout forms up one folder
