+++ src/routes/Servers/Servers.tsx
@@ -4,7 +4,7 @@ import { getAllLoadoutJson, serverKeyExists } from '@/api'
 import { useQuery } from '@tanstack/react-query'
 import { FirstTimeSetup } from './components/Forms/FirstTimeSetup/FirstTimeSetup'
 import { CreateLoadoutSheet } from './components/Forms/Loadouts/CreateLoadout/CreateLoadoutSheet'
-import { UploadLoadoutSheet } from './components/Forms/Loadouts/UploadLoadout/UploadLoadoutSheet'
+import { UploadLoadoutSheet } from './components/Forms/UploadLoadout/UploadLoadoutSheet'
 import { StartupSheet } from './components/Forms/Startup/StartupSheet'
 import { LoadoutContainer } from './components/Loadouts/LoadoutContainer'
 import { Button } from '@/components/ui/button'
