+++ src/routes/Servers/components/Forms/CreateLoadout/Setup/defaultServerConfig.ts
