+++ src/routes/Servers/components/Loadouts/Loadout.tsx
@@ -1,9 +1,8 @@
-import { Edit, Folder, Server, User } from 'lucide-react'
+import { Folder, Server, User } from 'lucide-react'
 import { openExplorerAtLoadout, playVUOnLocalServer, startServerLoadout } from '@/api'
 import { toast } from 'sonner'
 import { LoadoutJSON } from '@/config/config'
 import { ChooseAccountSheet } from '../Forms/AccountMultiSelect/AccountMultiSelectSheet'
-import { EditLoadoutSheet } from '../Forms/Loadouts/EditLoadout/EditLoadoutSheet'
 import { StartupSheet } from '../Forms/Startup/StartupSheet'
 import { ManageModsInServerSheet } from '../ManageModsInServerSheet/ManageModsInServerSheet'
 import { MaplistSheet } from '../Forms/Maplist/MaplistSheet'
